+++ history.md
@@ -7,6 +7,7 @@
 - Removed `rsyntaxtextarea-3.1.1.jar` from the externals. It was never used so far.
 - Class `mpmToolbox.gui.msmTree.MsmTree` has been expanded. It now provides its own instance of a `WebDockableFrame`. A simple getter method makes it easier to be used in class `mpmToolbox.gui.ProjectPane`. Thus, some code optimizations have been applied there.
 - New class `mpmToolbox.gui.mpmTree.MpmDockableFrame` has been added. This simplifies some code in class `mpmToolbox.gui.ProjectPane`.
+- Some first preparations for the audio analysis component.
 
 
 #### v0.1.4